+++ saleor/graphql/webhook/tests/mutations/test_webhook_create.py
@@ -1,3 +1,5 @@
+import json
+
 import graphene
 
 from .....webhook.error_codes import WebhookErrorCode
@@ -30,6 +32,7 @@ WEBHOOK_CREATE = """
 def test_webhook_create_by_app(app_api_client, permission_manage_orders):
     # given
     query = WEBHOOK_CREATE
+    custom_headers = {"X-Key": "Value", "Authorization-Key": "Value"}
     variables = {
         "input": {
             "name": "New integration",
@@ -38,6 +41,7 @@ def test_webhook_create_by_app(app_api_client, permission_manage_orders):
                 WebhookEventTypeAsyncEnum.ORDER_CREATED.name,
                 WebhookEventTypeAsyncEnum.ORDER_CREATED.name,
             ],
+            "customHeaders": json.dumps(custom_headers),
         }
     }
 
@@ -54,6 +58,7 @@ def test_webhook_create_by_app(app_api_client, permission_manage_orders):
     new_webhook = Webhook.objects.get()
     assert new_webhook.name == "New integration"
     assert new_webhook.target_url == "https://www.example.com"
+    assert new_webhook.custom_headers == custom_headers
     events = new_webhook.events.all()
     assert len(events) == 1
     assert events[0].event_type == WebhookEventTypeAsyncEnum.ORDER_CREATED.value
@@ -286,3 +291,31 @@ def test_webhook_create_inherit_events_from_query(
         data["webhook"]["syncEvents"][0]["eventType"]
         == WebhookEventTypeSyncEnum.PAYMENT_LIST_GATEWAYS.name
     )
+
+
+def test_webhook_create_invalid_custom_headers(app_api_client):
+    # given
+    query = WEBHOOK_CREATE
+    custom_headers = {"DisallowedKey": "Value"}
+    variables = {
+        "input": {
+            "name": "New integration",
+            "targetUrl": "https://www.example.com",
+            "customHeaders": json.dumps(custom_headers),
+        }
+    }
+
+    # when
+    response = app_api_client.post_graphql(query, variables=variables)
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["webhookCreate"]
+    assert not data["webhook"]
+    error = data["errors"][0]
+    assert error["field"] == "customHeaders"
+    assert (
+        error["message"] == '"DisallowedKey" does not match allowed key pattern: '
+        '"X-*" or "Authorization*".'
+    )
+    assert error["code"] == WebhookErrorCode.INVALID_CUSTOM_HEADERS.name