+++ saleor/graphql/webhook/tests/mutations/test_webhook_update.py
@@ -1,5 +1,8 @@
+import json
+
 import graphene
 
+from ....core.enums import WebhookErrorCode
 from ....tests.utils import assert_no_permission, get_graphql_content
 from ...enums import WebhookEventTypeAsyncEnum
 
@@ -9,6 +12,7 @@ WEBHOOK_UPDATE = """
         errors {
           field
           message
+          code
         }
         webhook {
           syncEvents {
@@ -18,6 +22,7 @@ WEBHOOK_UPDATE = """
             eventType
           }
           isActive
+          customHeaders
         }
       }
     }
@@ -46,6 +51,7 @@ def test_webhook_update_by_staff(staff_api_client, webhook, permission_manage_ap
     # given
     query = WEBHOOK_UPDATE
     webhook_id = graphene.Node.to_global_id("Webhook", webhook.pk)
+    custom_headers = {"X-Key": "Value", "Authorization-Key": "Value"}
     variables = {
         "id": webhook_id,
         "input": {
@@ -54,6 +60,7 @@ def test_webhook_update_by_staff(staff_api_client, webhook, permission_manage_ap
                 WebhookEventTypeAsyncEnum.CUSTOMER_CREATED.name,
             ],
             "isActive": False,
+            "customHeaders": json.dumps(custom_headers),
         },
     }
     staff_api_client.user.user_permissions.add(permission_manage_apps)
@@ -65,6 +72,7 @@ def test_webhook_update_by_staff(staff_api_client, webhook, permission_manage_ap
 
     # then
     assert webhook.is_active is False
+    assert webhook.custom_headers == custom_headers
     events = webhook.events.all()
     assert len(events) == 1
     assert events[0].event_type == WebhookEventTypeAsyncEnum.CUSTOMER_CREATED.value
@@ -76,6 +84,7 @@ def test_webhook_update_by_staff(staff_api_client, webhook, permission_manage_ap
         == WebhookEventTypeAsyncEnum.CUSTOMER_CREATED.name
     )
     assert data["webhook"]["isActive"] is False
+    assert data["webhook"]["customHeaders"] == json.dumps(custom_headers)
 
 
 def test_webhook_update_by_staff_without_permission(staff_api_client, app, webhook):
@@ -131,3 +140,36 @@ def test_webhook_update_inherit_events_from_query(
     events = webhook.events.all()
     assert len(events) == 1
     assert WebhookEventTypeAsyncEnum.ORDER_UPDATED.value == events[0].event_type
+
+
+def test_webhook_update_invalid_custom_headers(
+    staff_api_client,
+    webhook,
+    permission_manage_apps,
+):
+    # given
+    query = WEBHOOK_UPDATE
+    custom_headers = {"DisallowedKey": "Value"}
+    webhook_id = graphene.Node.to_global_id("Webhook", webhook.pk)
+    staff_api_client.user.user_permissions.add(permission_manage_apps)
+    variables = {
+        "id": webhook_id,
+        "input": {
+            "customHeaders": json.dumps(custom_headers),
+        },
+    }
+
+    # when
+    response = staff_api_client.post_graphql(query, variables=variables)
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["webhookUpdate"]
+    assert not data["webhook"]
+    error = data["errors"][0]
+    assert error["field"] == "customHeaders"
+    assert (
+        error["message"] == '"DisallowedKey" does not match allowed key pattern: '
+        '"X-*" or "Authorization*".'
+    )
+    assert error["code"] == WebhookErrorCode.INVALID_CUSTOM_HEADERS.name